Working

+++ src/Controller/UserAddressController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\UserAddress;
 use App\Form\UserAddressType;
 use App\Repository\UserAddressRepository;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -12,6 +13,7 @@ use Symfony\Component\Routing\Annotation\Route;
 
 /**
  * @Route("/user/address")
+ * @IsGranted("ROLE_USER")
  */
 class UserAddressController extends AbstractController
 {
@@ -20,8 +22,15 @@ class UserAddressController extends AbstractController
      */
     public function index(UserAddressRepository $userAddressRepository): Response
     {
+        $user = $this->getUser();
+
+        $userAddress = $userAddressRepository->findBy([
+            'user'   => $user,
+            'status' => UserAddress::STATUS_ACTIVE
+        ]);
+
         return $this->render('user_address/index.html.twig', [
-            'user_addresses' => $userAddressRepository->findAll(),
+            'user_addresses' => $userAddress,
         ]);
     }
 
@@ -49,20 +58,12 @@ class UserAddressController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="user_address_show", methods={"GET"})
-     */
-    public function show(UserAddress $userAddress): Response
-    {
-        return $this->render('user_address/show.html.twig', [
-            'user_address' => $userAddress,
-        ]);
-    }
-
-    /**
      * @Route("/{id}/edit", name="user_address_edit", methods={"GET","POST"})
      */
     public function edit(Request $request, UserAddress $userAddress): Response
     {
+        $this->denyAccessUnlessGranted('USER_ADDRESS_MANAGE', $userAddress);
+
         $form = $this->createForm(UserAddressType::class, $userAddress);
         $form->handleRequest($request);
 
@@ -85,6 +86,8 @@ class UserAddressController extends AbstractController
      */
     public function delete(Request $request, UserAddress $userAddress): Response
     {
+        $this->denyAccessUnlessGranted('USER_ADDRESS_MANAGE', $userAddress);
+
         if ($this->isCsrfTokenValid('delete'.$userAddress->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($userAddress);
